--- examples/pump-controller-app/pump-controller-common/gen/CHIPClustersObjc.h
@@ -1,89 +1,0 @@
-/*
- *
- *    Copyright (c) 2021 Project CHIP Authors
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-// THIS FILE IS GENERATED BY ZAP
-
-#ifndef CHIP_CLUSTERS_H
-#define CHIP_CLUSTERS_H
-
-#import <Foundation/Foundation.h>
-
-@class CHIPDevice;
-
-typedef void (^ResponseHandler)(NSError * _Nullable error, NSDictionary * _Nullable values);
-
-NS_ASSUME_NONNULL_BEGIN
-
-/**
- * CHIPCluster
- *    This is the base class for clusters.
- */
-@interface CHIPCluster : NSObject
-
-- (nullable instancetype)initWithDevice:(CHIPDevice *)device
-                               endpoint:(uint8_t)endpoint
-                                  queue:(dispatch_queue_t)queue NS_DESIGNATED_INITIALIZER;
-- (instancetype)init NS_UNAVAILABLE;
-+ (instancetype)new NS_UNAVAILABLE;
-
-@end
-
-/**
- * Cluster On/off
- *
- */
-@interface CHIPOnOff : CHIPCluster
-
-- (void)off:(ResponseHandler)responseHandler;
-- (void)on:(ResponseHandler)responseHandler;
-- (void)toggle:(ResponseHandler)responseHandler;
-
-- (void)readAttributeOnOffWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)configureAttributeOnOffWithMinInterval:(uint16_t)minInterval
-                                   maxInterval:(uint16_t)maxInterval
-                               responseHandler:(ResponseHandler)responseHandler;
-- (void)reportAttributeOnOffWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeClusterRevisionWithResponseHandler:(ResponseHandler)responseHandler;
-
-@end
-
-/**
- * Cluster Pump Configuration and Control
- *
- */
-@interface CHIPPumpConfigurationAndControl : CHIPCluster
-
-- (void)readAttributeMaxPressureWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeMaxSpeedWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeMaxFlowWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeEffectiveOperationModeWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeEffectiveControlModeWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeCapacityWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)configureAttributeCapacityWithMinInterval:(uint16_t)minInterval
-                                      maxInterval:(uint16_t)maxInterval
-                                           change:(int16_t)change
-                                  responseHandler:(ResponseHandler)responseHandler;
-- (void)reportAttributeCapacityWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeOperationModeWithResponseHandler:(ResponseHandler)responseHandler;
-- (void)writeAttributeOperationModeWithValue:(uint8_t)value responseHandler:(ResponseHandler)responseHandler;
-- (void)readAttributeClusterRevisionWithResponseHandler:(ResponseHandler)responseHandler;
-
-@end
-
-NS_ASSUME_NONNULL_END
-
-#endif /* CHIP_CLUSTERS_H */
