+++ src/darwin/Framework/CHIP/templates/CHIPClustersObjc.zapt
