+++ src/darwin/Framework/CHIP/templates/templates.json
@@ -9,8 +9,23 @@
         "helper.js"
     ],
     "override": "../../../../../src/app/zap-templates/common/override.js",
-    "partials": [],
+    "partials": [
+        {
+            "name": "header",
+            "path": "../../../../../src/app/zap-templates/partials/header.zapt"
+        }
+    ],
     "templates": [
+        {
+            "path": "CHIPClustersObjc.zapt",
+            "name": "Objc ZCL API Header",
+            "output": "src/darwin/Framework/CHIP/gen/CHIPClustersObjc.h"
+        },
+        {
+            "path": "CHIPClustersObjc-src.zapt",
+            "name": "Objc ZCL API",
+            "output": "src/darwin/Framework/CHIP/gen/CHIPClustersObjc.mm"
+        },
         {
             "path": "clusters-tests.zapt",
             "name": "Cluster Tests",
